pyproject.toml
[tool.poetry]
name = "peopeople"
version = "0.1.0"
description = ""
authors = ["DraksPlay <draksplay@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
aiohttp = "^3.9.3"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
